refactor: compress
